cardhtml as typescript class
